feat: make the inclusion of params in snippets explicit
Previously, every parameter was represented in every code snippet
whether or not it added value. This policy leads to messy looking
examples that may be non-sensical or overly complicated or that seem
to emphasize obscure parameters and use cases.

With this change, parameters are included in a code snippet if and only
if they are needed or intended.

The new rules are the following:

- A required parameter will be included in the code snippet. The app
  looks for a value in the values passed to getParameterValues. Next
  it checks in order param.example, param.examples,
  param.schema.example and finally param.default.
- Path parameters are always required and therefore always present
- A non-required parameter will only be included if a value is
  supplied in the values array passed to getParameterValues or if
  there is an example in `param.examples` that includes the vendor
  extension `x-use-in-code-snippets: true`.
- It's not possible for param.example or param.default or
  param.schema.example values to show up in code snippet if the
  parameter is not required.