chore: update dependencies
The biggest change is in httpsnippet which validates urls
are valid before creating a snippet. Many of the specs used
path variables (for example `{location-id}`) which when left
in place render the url invalid. To work around this, every
path parameter in 2.x spec needs to have a `default` value and
every 3.0 spec needs to have example or examples or default
in the schema.

All of the operations used in tests were updated in this way.